feat (main): main method checks if there are available ingredients to make the desired size of sandwich.
customer_order receives what size of sandwich the customer wants.
Cost is determined based off what it takes to make the S/M/L sandwich.
recipes in data.py contains the information that includes the amount of resources available as well as how much each sandwich costs.

If there are enough resources available to make the sandwich, the total is printed out.