feat(cashier.py): transaction _result function has been completed.
Coins parameter is checked to either be greater than or equal to cost parameter. The coins are subtracted by the cost and rounded to the second decimal point. This number is then assigned to change.

If the change is greater than 0, a message is printed showing the change due back to the customer.

If the change is equal to 0, a message is printed showing there is not any change due back to the customer.

If the coins are less than the cost, a message is printed showing the customer did not input the enough money.